moved `scrollToBottomOfAnnotationStep()` into utils file
